+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/ActivityContext.kt
@@ -1,6 +1,7 @@
 package io.github.pandier.intellijdiscordrp.activity
 
 import com.intellij.openapi.application.ApplicationInfo
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ApplicationNamesInfo
 import com.intellij.openapi.application.ReadAction
 import com.intellij.openapi.editor.Editor
@@ -27,33 +28,36 @@ class ActivityFileContext(
     val line: Int?,
     val lineCount: Int?,
     val length: Long?,
+    val start: Instant,
 )
 
 class ActivityContext(
-    val start: Instant,
     val appName: String,
     val appFullName: String,
     val appVersion: String,
+    val appStart: Instant,
     val project: WeakReference<Project>,
     val projectName: String,
+    val projectStart: Instant,
     val file: ActivityFileContext? = null,
 ) {
     companion object Factory {
         fun create(
             project: Project,
             file: VirtualFile? = null,
-            editor: Editor? = null,
-            start: Instant = TimeTrackingService.getInstance(project).start,
+            editor: Editor? = null
         ): ActivityContext {
+            val timeTrackingService = TimeTrackingService.getInstance()
             val appInfo = ApplicationInfo.getInstance()
             val appNames = ApplicationNamesInfo.getInstance()
             return ActivityContext(
-                start = start,
                 appName = appNames.fullProductName,
                 appFullName = appNames.fullProductNameWithEdition,
                 appVersion = appInfo.fullVersion,
-                projectName = project.name,
+                appStart = timeTrackingService.getOrInit(ApplicationManager.getApplication()),
                 project = WeakReference(project),
+                projectName = project.name,
+                projectStart = timeTrackingService.getOrInit(project),
                 file = file?.let {
                     val contentRoot = ReadAction.compute<VirtualFile?, Exception> {
                         ProjectFileIndex.getInstance(project).getContentRootForFile(it)
@@ -69,6 +73,7 @@ class ActivityContext(
                         line = editor?.caretModel?.logicalPosition?.line?.plus(1),
                         lineCount = editor?.document?.lineCount?.let { max(it, 1) },
                         length = it.length,
+                        start = timeTrackingService.getOrInit(it),
                     )
                 }
             )